# feat(iOS): allow eager initialization of RCTRootViewFactory