--- a/packages/react-native/Libraries/AppDelegate/RCTRootViewFactory.h
+++ b/packages/react-native/Libraries/AppDelegate/RCTRootViewFactory.h
@@ -219,7 +219,8 @@ typedef void (^RCTLoadSourceForBridgeBlock)(RCTBridge *bridge, RCTSourceLoadBloc
  * 
  * @parameter: launchOptions  - a dictionary with a set of options.
  */
-- (void)initializeReactHostWithLaunchOptions:(NSDictionary *__nullable)launchOptions
+- (void)initializeReactHostWithLaunchOptions:(NSDictionary *__nullable)launchOptions;
+
 - (RCTHost *)createReactHost:(NSDictionary *__nullable)launchOptions;
 
 @end